Entry migration and db configuration + modified models

+++ Base of Taste/Base of Taste/Models/SkladnikDoPrzepis.cs
@@ -7,7 +7,7 @@ namespace BazaSmakuAPI.Models
         public int ID { get; set; }
         public int Skladnik_ID { get; set; }
         public int Przepis_ID { get; set; }
-        public int Ilosc { get; set; }
+        public decimal Ilosc { get; set; }
         public string Jednostka_ID { get; set; }
 
     }
